feat(auth): Refactorizar el callback de OAuth para mejorar el manejo de errores y el logging; actualizar la obtención del email para usar UnauthorizedError
