+++ src/services/authService.ts
@@ -3,7 +3,8 @@ import { oauth2Client } from "../config/googleOAuth";
 import userService from "./userService";
 import jwtService from "./jwtService";
 import { ensureOAuthAccess } from "../config/oAuthAccess";
-import { auditService } from "./auditService";
+import auditService from "./auditService";
+import { UnauthorizedError } from "../errors/AppError";
 
 const oauth2 = google.oauth2("v2");
 
@@ -23,44 +24,50 @@ class AuthService {
     refreshToken: string;
     redirectUrl: string;
   }> {
-    const { tokens } = await oauth2Client.getToken(code);
-    oauth2Client.setCredentials(tokens);
+    try {
+      const { tokens } = await oauth2Client.getToken(code);
+      oauth2Client.setCredentials(tokens);
 
-    const { data: profile } = await oauth2.userinfo.get({ auth: oauth2Client });
-    const email = profile.email?.toLowerCase();
+      const { data: profile } = await oauth2.userinfo.get({ auth: oauth2Client });
+      const email = profile.email?.toLowerCase();
 
-    if (!email) {
-      throw new Error("No se pudo obtener el email de Google");
+      if (!email) {
+        throw new UnauthorizedError("No se pudo obtener el email de Google");
+      }
+
+      // Donde se verifica si el usuario tiene acceso permitido, por ahora, sólo la lista de mails permitidos. 
+      ensureOAuthAccess(email);
+
+      const role = userService.determineUserRole(email);
+      const user = await userService.upsertUser({
+        email,
+        name: profile.name ?? "",
+        role,
+      });
+
+      const accessToken = jwtService.generateAccessToken(user.id, user.email, user.role);
+      const refreshToken = jwtService.generateRefreshToken(user.id);
+
+      const frontendURL = process.env.FRONTEND_URL!;
+      const queryParams = new URLSearchParams({
+        success: "true",
+        role: user.role,
+      });
+
+      // Registro de login exitoso en auditoría 
+      auditService.recordLogin(user.email, user.name).catch((err) => {
+        console.error('[AuthService][audit] recordLogin failed:', err);
+      });
+
+      return {
+        accessToken,
+        refreshToken,
+        redirectUrl: `${frontendURL}/auth/callback?${queryParams.toString()}`,
+      };
+    } catch (error) {
+      console.error('[AuthService] Error en OAuth callback:', error);
+      throw error;
     }
-
-    // Donde se verifica si el usuario tiene acceso permitido, por ahora, sólo la lista de mails permitidos. 
-    ensureOAuthAccess(email);
-
-    const role = userService.determineUserRole(email);
-    const user = await userService.upsertUser({
-      email,
-      name: profile.name ?? "",
-      role,
-    });
-
-    const accessToken = jwtService.generateAccessToken(user.id, user.email, user.role);
-    const refreshToken = jwtService.generateRefreshToken(user.id);
-
-    const frontendURL = process.env.FRONTEND_URL!;
-    const queryParams = new URLSearchParams({
-      success: "true",
-    });
-
-    // Registro de login exitoso en auditoría (no bloquea el flujo de autenticación)
-    auditService.recordLogin(user.email).catch((err) => {
-      console.error('[AuthService][audit] recordLogin failed:', err);
-    });
-
-    return {
-      accessToken,
-      refreshToken,
-      redirectUrl: `${frontendURL}/auth/callback?${queryParams.toString()}`,
-    };
   }
 }
